Passed the first stage of the website html of odin-recipes.
As a result of the training, a five-page website was created.
All pages are created with HTML markup language. Added five images
in JPG format to the pages as illustrations of ready meals and five
images in PNG format as page backgrounds. All conditions of the task
are fulfilled: Four child pages are attached to the main page by
relative links in the form of an unnumbered list. There are also
unnumbered and numbered lists with listing of ingredients and cooking
periods and titles of different sizes. In the future, it is planned to
refine the site as the CSS style language and the JC script language
are studied.